// src/resources/functions/secrets/project.ts
// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

import { APIResource } from '../../../core/resource';
import { APIPromise } from '../../../core/api-promise';
import { buildHeaders } from '../../../internal/headers';
import { RequestOptions } from '../../../internal/request-options';
import { path } from '../../../internal/utils/path';

export class BaseProject extends APIResource {
  static override readonly _key: readonly ['functions', 'secrets', 'project'] = Object.freeze(['functions', 'secrets', 'project'] as const)

  retrieve(projectID: string, options?: RequestOptions): APIPromise<void> {
    return this._client.get(path`/functions/secrets/project/${projectID}`, { ...options, headers: buildHeaders([{Accept: '*/*'}, options?.headers]) });
  }
}
export class Project extends BaseProject {

}